Eliminate compiler warnings for implicit int64->32 conversion.  Most of these weren't bugs, but in some cases, it generalizes the code a bit in ways that should be nice for future flexibility. Change: 132773101

+++ tensorflow/core/lib/strings/numbers.cc
@@ -279,7 +279,7 @@ bool safe_strto32(StringPiece str, int32* value) {
 
   if (!str.empty()) return false;
 
-  *value = result * sign;
+  *value = static_cast<int32>(result * sign);
   return true;
 }
 
@@ -299,7 +299,7 @@ bool safe_strtou32(StringPiece str, uint32* value) {
   SkipSpaces(&str);
   if (!str.empty()) return false;
 
-  *value = result;
+  *value = static_cast<uint32>(result);
   return true;
 }
